Seperated data for training
This data is split so that IJ is the training set and K is the test set. IJ contains two experimental replicates with input copies of 30-10000 while K contains a single experiment with input copies of 30-10000. This negates the need for generating a train test split while maintaining the ability to test across the entire experimental range.